requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
distlib==0.4.0
Django==4.2.23
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
gensim==4.3.3
gevent==24.2.1
greenlet==3.1.1
huggingface-hub==0.33.4
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
importlib_resources==6.4.5
inflection==0.5.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
kombu==5.5.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==6.1.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
rpds-py==0.20.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==3.2.1
six==1.17.0
smart_open==7.3.0.post1
sqlparse==0.5.3
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
python-barcode==0.16.0
tw-invoice == 2023.9.6
cryptography>=41.0 